Remove null credentials fixed.

+++ example-app/src/main/java/org/github/sipuada/plugins/android/audio/example/presenter/SipuadaService.java
@@ -13,7 +13,7 @@ import android.os.Process;
 import android.support.annotation.Nullable;
 import android.util.Log;
 
-import com.activeandroid.query.Select;
+import com.activeandroid.query.;
 import com.google.common.eventbus.DeadEvent;
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
@@ -657,7 +657,7 @@ public class SipuadaService extends Service {
         }
         SipuadaUserCredentials self = new Select().from(SipuadaUserCredentials.class)
                 .where("Username = ? AND PrimaryHost = ?", username, primaryHost).executeSingle();
-        self.delete();
+        if(self != null) self.delete();
     }
 
     abstract class SipuadaServiceListener implements SipuadaApi.SipuadaListener {
